Add exceptions in all services.
Add Aborted and ErrorConnection for cross-service communication purposes.
Document DB-Service contracts (in thrift interface files).

+++ db-services/db-diagram-service/src/main/java/com/qreal/wmp/db/diagram/dao/DiagramDaoImpl.java
@@ -1,5 +1,7 @@
 package com.qreal.wmp.db.diagram.dao;
 
+import com.qreal.wmp.db.diagram.exceptions.Aborted;
+import com.qreal.wmp.db.diagram.exceptions.NotFound;
 import com.qreal.wmp.db.diagram.model.Diagram;
 import com.qreal.wmp.db.diagram.model.Folder;
 import org.hibernate.Session;
@@ -38,11 +40,15 @@ class DiagramDaoImpl implements DiagramDao {
      * @param folderId id of folder to find
      */
     @Override
-    public Folder getFolder(Long folderId) {
+    @NotNull
+    public Folder getFolder(Long folderId) throws NotFound {
         logger.trace("getFolder method called with parameters: folderID = {}", folderId);
         Session session = sessionFactory.getCurrentSession();
         Folder result = (Folder) session.get(Folder.class, folderId);
         logger.trace("getFolder method extracted result from session and return it");
+        if (result == null) {
+            throw new NotFound(String.valueOf(folderId), "Folder with specified Id not found.");
+        }
         return result;
     }
 
@@ -53,20 +59,19 @@ class DiagramDaoImpl implements DiagramDao {
      * @param folderId id of folder to save diagram in
      */
     @Override
-    public Long saveDiagram(@NotNull Diagram diagram, Long folderId) {
+    public Long saveDiagram(@NotNull Diagram diagram, Long folderId) throws Aborted {
         logger.trace("saveDiagram method called with parameters: diagram = {}, folderID = {}", diagram.getName(),
                 folderId);
         Session session = sessionFactory.getCurrentSession();
-        Folder folder = getFolder(folderId);
-        if (folder == null) {
+        Folder folder = null;
+        List<Diagram> diagrams = null;
+        try {
+            folder = getFolder(folderId);
+            diagrams = folder.getDiagrams();
+        } catch (NotFound e) {
             logger.error("Got null folder object for folder id {}.", folderId);
-            return 0L; // FIXME
-        }
-
-        List<Diagram> diagrams = folder.getDiagrams();
-        if (diagrams == null) {
-            logger.error("Got empty diagrams list for folder id {}.", folderId);
-            return 0L; // FIXME
+            throw new Aborted("Folder to save diagram into not found", "saveDiagram safely aborted",
+                    DiagramDaoImpl.class.getName(), e);
         }
         diagrams.add(diagram);
         session.update(folder);
@@ -81,14 +86,16 @@ class DiagramDaoImpl implements DiagramDao {
      * @param diagramId id of diagram to find
      */
     @Override
-    public Diagram openDiagram(Long diagramId) {
+    @NotNull
+    public Diagram openDiagram(Long diagramId) throws NotFound {
         logger.trace("openDiagram method called with parameters: id = {}", diagramId);
         Session session = sessionFactory.getCurrentSession();
-        List<Diagram> diagrams = session.createQuery("from Diagram where id=:diagramId").
-                setParameter("diagramId", diagramId).list();
-        logger.trace("openDiagram method extracted list of results from session with {} elements. First will be " +
-                "returned.", diagrams.size());
-        return diagrams.stream().findFirst().orElse(null);
+        Diagram diagram = (Diagram) session.get(Diagram.class, diagramId);
+        if (diagram == null) {
+            throw new NotFound(String.valueOf(diagramId), "Diagram with specified Id not found.");
+        }
+        logger.trace("openDiagram method extracted diagram");
+        return diagram;
     }
 
     /** Checks whether diagram with specified id exists.*/
@@ -96,22 +103,23 @@ class DiagramDaoImpl implements DiagramDao {
     public boolean isExistsDiagram(Long diagramId) {
         logger.trace("isExistsDiagram method called with parameters: id = {}", diagramId);
         Session session = sessionFactory.getCurrentSession();
-        List<Diagram> diagrams = session.createQuery("from Diagram where id=:diagramId").
-                setParameter("diagramId", diagramId).list();
-        logger.trace("isExistsDiagram method extracted list of results from session with {} elements.",
-                diagrams.size());
-        return !diagrams.isEmpty();
+        Diagram diagram = (Diagram) session.get(Diagram.class, diagramId);
+        return diagram != null;
     }
 
     /**
      * Rewrites diagram at local DB using Hibernate ORM.
      *
-     * @param diagram diagram to rewrite (<code>diagram.id</code> must be set correctly).
+     * @param diagram diagram to rewrite (Id must be set correctly).
      */
     @Override
-    public void rewriteDiagram(@NotNull Diagram diagram) {
+    public void rewriteDiagram(@NotNull Diagram diagram) throws Aborted {
         logger.trace("rewriteDiagram method called with parameters: diagram = {}", diagram.getName());
         Session session = sessionFactory.getCurrentSession();
+        if (!isExistsDiagram(diagram.getId())) {
+            throw new Aborted("Diagram with specified Id doesn't exist. Use save instead.", "rewriteDiagram safely " +
+                    "aborted.", DiagramDaoImpl.class.getName());
+        }
         session.merge(diagram);
         logger.trace("rewriteDiagram method edited diagram {}", diagram.getName());
     }
@@ -123,9 +131,13 @@ class DiagramDaoImpl implements DiagramDao {
      * @param diagramId id of diagram to delete
      */
     @Override
-    public void deleteDiagram(Long diagramId) {
+    public void deleteDiagram(Long diagramId) throws Aborted {
         logger.trace("deleteDiagram method called with parameters: id = {}", diagramId);
         Session session = sessionFactory.getCurrentSession();
+        if (!isExistsDiagram(diagramId)) {
+            throw new Aborted("Diagram with specified Id doesn't exist.", "deleteDiagram safely aborted.",
+                    DiagramDaoImpl.class.getName());
+        }
         Diagram diagram = (Diagram) session.get(Diagram.class, diagramId);
         session.delete(diagram);
         logger.trace("deleteDiagram method deleted diagram with id {}", diagramId);
@@ -151,12 +163,9 @@ class DiagramDaoImpl implements DiagramDao {
     public boolean isExistsFolder(Long folderId) {
         logger.trace("isExistsFolder method called with parameters: id = {}", folderId);
         Session session = sessionFactory.getCurrentSession();
-        List<Folder> folders = session.createQuery("from Folder where id=:folderId").
-                setParameter("folderId", folderId).list();
-        logger.trace("isExistsFolder method extracted list of results from session with {} elements.", folders.size());
-        return !folders.isEmpty();    }
-
-
+        Folder folder = (Folder) session.get(Folder.class, folderId);
+        return folder != null;
+    }
 
     /**
      * Deletes folder from local DB using Hibernate ORM.
@@ -164,9 +173,13 @@ class DiagramDaoImpl implements DiagramDao {
      * @param folderId id of folder to delete
      */
     @Override
-    public void deleteFolder(Long folderId) {
+    public void deleteFolder(Long folderId) throws Aborted {
         logger.trace("deleteFolder method called with parameters: folderId = {}", folderId);
         Session session = sessionFactory.getCurrentSession();
+        if (!isExistsDiagram(folderId)) {
+            throw new Aborted("Folder with specified Id doesn't exist.", "deleteFolder safely aborted.",
+                    DiagramDaoImpl.class.getName());
+        }
         Folder folder = (Folder) session.get(Folder.class, folderId);
         session.delete(folder);
         logger.trace("deleteFolder method deleted folder with id {}", folderId);
@@ -178,7 +191,8 @@ class DiagramDaoImpl implements DiagramDao {
      * @param userName name of user which root folder seeking
      */
     @Override
-    public Folder getFolderTree(String userName) {
+    @NotNull
+    public Folder getFolderTree(String userName) throws NotFound {
         logger.trace("getFolderTree method called with parametrs: userName = {}", userName);
         Session session = sessionFactory.getCurrentSession();
 
@@ -186,6 +200,9 @@ class DiagramDaoImpl implements DiagramDao {
                 "userName=:userName").setParameter("folderName", "root").setParameter("userName", userName).list();
         logger.trace("getFolderTree method extracted list of results from session with {} elements. First will be " +
                 "returned.", rootFolders.size());
-        return rootFolders.stream().findFirst().orElse(null);
+        if (rootFolders.isEmpty()) {
+            throw new NotFound(String.valueOf(userName), "FolderTree for user with specified username not found.");
+        }
+        return rootFolders.get(0);
     }
 }
